+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 UNRELEASED
 ----------
 
+* [ [#1664](https://github.com/digitalfabrik/integreat-cms/issues/1664) ] Fix media library and content edit lock in sbs view
+* [ [#1660](https://github.com/digitalfabrik/integreat-cms/issues/1660) ] Fix moving pages to the root level from the page form
+* [ [#1566](https://github.com/digitalfabrik/integreat-cms/issues/1566) ] Manage organizations per region
+* [ [#1440](https://github.com/digitalfabrik/integreat-cms/issues/1440) ] Add missing word count to translation coverage report
+* [ [#1596](https://github.com/digitalfabrik/integreat-cms/issues/1596) ] Ignore default language in translation report
+* [ [#1607](https://github.com/digitalfabrik/integreat-cms/issues/1607) ] Count words of source translation in translation report
+* [ [#1674](https://github.com/digitalfabrik/integreat-cms/issues/1674) ] Improve language tree node creation
+* [ [#1591](https://github.com/digitalfabrik/integreat-cms/issues/1591) ] Add bulk actions for language tree nodes
+* [ [#1673](https://github.com/digitalfabrik/integreat-cms/issues/1673) ] Add possibility to reject pending reviews and discard auto saves
+* [ [#1640](https://github.com/digitalfabrik/integreat-cms/issues/1640) ] Update author when restoring old versions
+* [ [#1583](https://github.com/digitalfabrik/integreat-cms/issues/1583) ] Only enable the chat for staff and management roles
+* [ [#1647](https://github.com/digitalfabrik/integreat-cms/issues/1647) ] Give service team delete permissions for regions and push notifications
+
 
 2022.8.3
 --------